Converted package bcec version BLP_LIB_BDE_BCE_2.5.0 from clearcase

+++ groups/bce/bcec/bcec_multipriorityqueue.h
@@ -1064,7 +1064,7 @@ template <typename TYPE>
 inline
 int bcec_MultipriorityQueue<TYPE>::numPriorities() const
 {
-    return d_heads.size();
+    return static_cast<int>(d_heads.size());
 }
 
 template <typename TYPE>
